Different relative paths to test data.
For some reason, on Travis CI, the paths to testdata needs to be
prepended with "..".  This is believed to be a difference between
Windows and Linux, though it could be a difference between Qt 5.0
and Qt 5.3.

Either way, its not important - just for unit tests anyway, so lets
just support both.

This may be revisited when I start the Mac OS builds.