Handle zero-length length-delimited values
The google_message1.dat sample message includes a length-delimited
field value with a length of 0 (indeed, its the very first field,
with tag 1). This, apparently, is allowed (I see no reason for it not
to be), but was being interpreted as a failure when
Message::readLengthDelimitedValue returned an y empty QByteArray.

This fix wraps the QByteArray result in a QVariant (which the caller
did invariably anyway), so that failure can be distinguished from valid
empty results.

Included relevant test.