Use split-time in HRM [IntTimes] section (#20)
This is an example where the HRM2 v1.4 documentation is ambiguous.
However, AR67 (from the Polar forums) tells me that these are meant to
be "split times", which appears to mean that they time-stamp the point
at which the lap was "taken".

The protobuf laps data includes a "split-time" field that appears to
match with what AR67 reports, so let's go with that for now :)

See http://forum.polar.fi/showthread.php?p=85839#post85839 for details.