Apply the Cluetrust GPXData 1.0 schema (#29)
The previous implementation was not actually validating anything, since
the Cluetrust GPXData XSD imports the GPX xsd, which allows anything to
be added to the GPX `extensions` element(s), thus passing anything and
everything.

Instead (or in addition), this change loops through all extensions, and
validates them against the Cluetrust GPXData XSD explicitly.