Move the hook files to the root of their own branch
This is part of a fundamental restructure of the Bipolar project code.
In this step, the hook code is being moved from master hook/qt to its
own hook branch.

This step / move is appropriate, since the hook follows a completely
different build / release pattern to the rest of the project, sinc it
is a patch to a specific historic version of Qt that matches FlowSync.