Unset the `VERSION` qmake variable for MinGW builds
Unfortunately, this is necessary due to a limitation in MinGW's
`windres` Windows resource compiler, that means it cannot handle
defines containing spaces. The presence of a `VERSION` variable
triggers qmake's automatic generation of a Windows resource file,
which we wish to avoid for MinGW, due to this bug.

Note, we don't really *need* spaces in our defines, however these unit
tests intentionally set the BUILD_USER to 'unit tests' to test / verify
exactly this sort of behaviour.

This also implies that anyone building the Bipolar.exe appliction,
should either not have spaces in their username, or manually override
the BUILD_USER value, or use MSVC not MinGW.