Include test data from @pueroman (#57)
This includes data from a Stages Power meter, which the V800 only
records on one side, resulting in Bipolar reporting incorrect power
values (issue #57).

This data will soon be used to test / prove the fixes to that issue.

Note, this is a verbatim copy of a sub-set of data provided by
@pueroman - specifically, the `*-route` and `*-finalize` data files
have been omitted (neither are necessary for fixing issue #57).