Locate FlowSync via the Windows registry #63
Specifically, this looks for the registry keys that the Polar FlowSync
for Windows installer writes to allow for inclusion in Windows' add/
remove programs application.

This check is in addition to the existing `%ProgramFiles%` checks, and
takes precedence over them, to allow for non-standard, but otherwise
perfectly valid, FlowSync install locations.