Support Qt 5.6 / C++11
By adding a space (or really anything other than underscore) between
the space-string literal and the underscore that followed it, we
prevent the compiler from interpreting the pair as a user-defined
literal (a concept added in C++11).

There's no need to add a space *before* the space-string literal, but
that just made it look a little nicer.