Don't use Qt 5.7's QVariant::toString for floats (#69)
Qt 5.7 added `QLocale::FloatingPointShortest` (see qt/qtbase@726fed0),
and updated `QVariant` to use that (instead of the previous
implementation added in Qt 5.5) when converting floats and doubles to
string, resulting in slightly different output, and QCOMPARE unit test
failures.

In this change, we use the Qt 5.5 / 5.6 implementation because its at
least as accurate as 5.7+, and implementing a 5.7-compatible version
would be a major undertaking (needing to duplicate the third-party
double-conversion code Qt borrows from the V8 project).