Use metres-per-second for TCX speeds #74
The Polar speed data in in kilometres-per-hour, but TCX uses metres-per-
second (apparently; the relevant schemas are not explicit about it).

For completeness, the HRM output is already using the correct units
(that is, kilometres-per-hour multiplied by 10 or 128 depending on the
situation), and the GPX output does not include speed data at all. So
neither of those formats needing correcting.