Add more power balance test data #81
This data contains some negative power samples that result in invalid
(or at least nonsense) HRM output. Also, one sample results in a
negative total power (not just left or right), which causes the
generated TCX file to be invalid when the Garmin Activity Extension is
in play, as that extension requires that the power values are positive
integers.

So, this commit will fail unit testing like:

```
Error XSDError in .../test/release/test, at line 3, column 9: Content of element
{http://www.garmin.com/xmlschemas/ActivityExtension/v2}Watts does not match its
type definition: -57 is not valid according to xs:unsignedShort..
```

That will be fixed in the next commit; doing so separately so that the
total changes of the fix can easily distinguished from this large
commit of mostly data files.

Thanks to @remifreiwald for providing yet another great test data set.